privatize sortable procedures
