+++ router_test.go
@@ -18,14 +18,14 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package yarpc_test
+package yarpc
 
 import (
 	"context"
 	"fmt"
+	"sort"
 	"testing"
 
-	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/api/middleware"
 	"go.uber.org/yarpc/api/middleware/middlewaretest"
 	"go.uber.org/yarpc/api/transport"
@@ -39,7 +39,7 @@ func TestMapRouter(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
-	m := yarpc.NewMapRouter("myservice")
+	m := NewMapRouter("myservice")
 
 	foo := transporttest.NewMockUnaryHandler(mockCtrl)
 	bar := transporttest.NewMockUnaryHandler(mockCtrl)
@@ -106,7 +106,7 @@ func TestMapRouter_Procedures(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
-	m := yarpc.NewMapRouter("myservice")
+	m := NewMapRouter("myservice")
 
 	bar := transport.NewUnaryHandlerSpec(transporttest.NewMockUnaryHandler(mockCtrl))
 	foo := transport.NewUnaryHandlerSpec(transporttest.NewMockUnaryHandler(mockCtrl))
@@ -154,7 +154,7 @@ func TestMapRouter_Procedures(t *testing.T) {
 }
 
 func TestEmptyProcedureRegistration(t *testing.T) {
-	m := yarpc.NewMapRouter("test-service-name")
+	m := NewMapRouter("test-service-name")
 
 	procedures := []transport.Procedure{
 		{
@@ -169,7 +169,7 @@ func TestEmptyProcedureRegistration(t *testing.T) {
 }
 
 func TestAmbiguousProcedureRegistration(t *testing.T) {
-	m := yarpc.NewMapRouter("test-service-name")
+	m := NewMapRouter("test-service-name")
 
 	procedures := []transport.Procedure{
 		{
@@ -188,7 +188,7 @@ func TestAmbiguousProcedureRegistration(t *testing.T) {
 }
 
 func TestEncodingBeforeWildcardProcedureRegistration(t *testing.T) {
-	m := yarpc.NewMapRouter("test-service-name")
+	m := NewMapRouter("test-service-name")
 
 	procedures := []transport.Procedure{
 		{
@@ -208,7 +208,7 @@ func TestEncodingBeforeWildcardProcedureRegistration(t *testing.T) {
 }
 
 func TestWildcardBeforeEncodingProcedureRegistration(t *testing.T) {
-	m := yarpc.NewMapRouter("test-service-name")
+	m := NewMapRouter("test-service-name")
 
 	procedures := []transport.Procedure{
 		{
@@ -238,10 +238,72 @@ func IgnoreTestRouterWithMiddleware(t *testing.T) {
 	routerMiddleware := middlewaretest.NewMockRouter(mockCtrl)
 	routerMiddleware.EXPECT().Choose(ctx, req, gomock.Any()).Times(1).Return(expectedSpec, nil)
 
-	router := middleware.ApplyRouteTable(yarpc.NewMapRouter("service"), routerMiddleware)
+	router := middleware.ApplyRouteTable(NewMapRouter("service"), routerMiddleware)
 
 	actualSpec, err := router.Choose(ctx, req)
 
 	assert.Equal(t, expectedSpec, actualSpec, "handler spec returned from route table did not match")
 	assert.Nil(t, err)
+}
+
+func TestProcedureSort(t *testing.T) {
+	ps := sortableProcedures{
+		{
+			Service:  "moe",
+			Name:     "echo",
+			Encoding: "json",
+		},
+		{
+			Service: "moe",
+		},
+		{
+			Service: "larry",
+		},
+		{
+			Service: "moe",
+			Name:    "ping",
+		},
+		{
+			Service:  "moe",
+			Name:     "echo",
+			Encoding: "raw",
+		},
+		{
+			Service: "moe",
+			Name:    "poke",
+		},
+		{
+			Service: "curly",
+		},
+	}
+	sort.Sort(ps)
+	assert.Equal(t, sortableProcedures{
+		{
+			Service: "curly",
+		},
+		{
+			Service: "larry",
+		},
+		{
+			Service: "moe",
+		},
+		{
+			Service:  "moe",
+			Name:     "echo",
+			Encoding: "json",
+		},
+		{
+			Service:  "moe",
+			Name:     "echo",
+			Encoding: "raw",
+		},
+		{
+			Service: "moe",
+			Name:    "ping",
+		},
+		{
+			Service: "moe",
+			Name:    "poke",
+		},
+	}, ps, "should order procedures lexicographically on (service, procedure, encoding)")
 }
